integrated into new Revery_Native.Icon module
