--- src/Native/Badge.re
@@ -1,2 +1,0 @@
-external set: string => unit = "revery_setBadge";
-external hide: unit => unit = "revery_hideBadge";
