+++ src/Native/Revery_Native.re
@@ -1,6 +1,5 @@
 module Dialog = Dialog;
 module Icon = Icon;
 module Notification = Notification;
-module Badge = Badge;
 
 include Initialization;
